fix: module dependencies
