+++ src/cli.rs
@@ -1,6 +1,5 @@
 use clap::{Parser, Subcommand};
-use config::{Config, ConfigError, File};
-use serde::Deserialize;
+use config::ConfigError;
 
 #[derive(Parser)]
 #[command(name = "techscan")]
@@ -32,121 +31,8 @@ pub enum Commands {
     },
 }
 
-#[derive(Deserialize, Default)]
-pub struct AppConfig {
-    pub exclude: Option<Vec<String>>,
-    pub reporter: Option<String>,
-}
-
 impl Cli {
     pub fn new() -> Result<Self, ConfigError> {
         Ok(Self::parse())
-    }
-
-    pub fn load_config_file(&self, path: &str) -> Result<AppConfig, ConfigError> {
-        let config = Config::builder()
-            .add_source(File::with_name(path))
-            .build()?;
-
-        config.try_deserialize()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use clap::Parser;
-
-    fn create_test_cli() -> Cli {
-        Cli::parse_from(["techscan", "lang", "."])
-    }
-
-    mod load_config_file {
-        use super::*;
-
-        #[test]
-        fn test_load_json_file() {
-            let cli = create_test_cli();
-            let config_path = "tests/fixtures/cli/config/complete.json";
-            let config = cli.load_config_file(config_path).unwrap();
-
-            assert_eq!(
-                config.exclude,
-                Some(vec!["*.test.*".into(), "exclude_dir".into()])
-            );
-            assert_eq!(config.reporter, Some("json".into()));
-        }
-
-        #[test]
-        fn test_load_json5_file() {
-            let cli = create_test_cli();
-            let config_path = "tests/fixtures/cli/config/complete.json5";
-            let config = cli.load_config_file(config_path).unwrap();
-
-            assert_eq!(
-                config.exclude,
-                Some(vec!["*.test.*".into(), "exclude_dir".into()])
-            );
-            assert_eq!(config.reporter, Some("json".into()));
-        }
-
-        #[test]
-        fn test_load_toml_file() {
-            let cli = create_test_cli();
-            let config_path = "tests/fixtures/cli/config/complete.toml";
-            let config = cli.load_config_file(config_path).unwrap();
-
-            assert_eq!(
-                config.exclude,
-                Some(vec!["*.test.*".into(), "exclude_dir".into()])
-            );
-            assert_eq!(config.reporter, Some("json".into()));
-        }
-
-        #[test]
-        fn test_load_yaml_file() {
-            let cli = create_test_cli();
-            let config_path = "tests/fixtures/cli/config/complete.yaml";
-            let config = cli.load_config_file(config_path).unwrap();
-
-            assert_eq!(
-                config.exclude,
-                Some(vec!["*.test.*".into(), "exclude_dir".into()])
-            );
-            assert_eq!(config.reporter, Some("json".into()));
-        }
-
-        #[test]
-        fn test_load_yml_file() {
-            let cli = create_test_cli();
-            let config_path = "tests/fixtures/cli/config/complete.yml";
-            let config = cli.load_config_file(config_path).unwrap();
-
-            assert_eq!(
-                config.exclude,
-                Some(vec!["*.test.*".into(), "exclude_dir".into()])
-            );
-            assert_eq!(config.reporter, Some("json".into()));
-        }
-
-        #[test]
-        fn test_load_reporter_only() {
-            let cli = create_test_cli();
-            let config_path = "tests/fixtures/cli/config/reporter_only.yaml";
-            let config = cli.load_config_file(config_path).unwrap();
-
-            assert_eq!(
-                config.exclude,
-                Some(vec!["*.test.*".into(), "exclude_dir".into()])
-            );
-            assert_eq!(config.reporter, None);
-        }
-
-        #[test]
-        fn test_nonexistent_config_file() {
-            let cli = create_test_cli();
-            let result = cli.load_config_file("non_existent.json");
-            assert!(result.is_err());
-        }
     }
 }
