+++ src/entity/app_config.rs
@@ -1,0 +1,7 @@
+use serde::Deserialize;
+
+#[derive(Deserialize, Default)]
+pub struct AppConfig {
+    pub exclude: Option<Vec<String>>,
+    pub reporter: Option<String>,
+}
