+++ src/entity/mod.rs
@@ -1,8 +1,10 @@
+pub mod app_config;
 pub mod file;
 pub mod language;
 pub mod language_report;
 pub mod language_scanner_options;
 
+pub use app_config::AppConfig;
 pub use file::File;
 pub use language::Language;
 pub use language_report::{LanguageReport, LanguageReportItem};
